fix type of cache profiling information printing

+++ src/sbdi_cache.c
@@ -115,6 +115,6 @@ sbdi_error_t sbdi_bc_cache_blk(sbdi_bc_t *cache, uint32_t blk_idx,
 #ifdef SBDI_CACHE_PROFILE
 void sbdi_bc_print_stats(sbdi_bc_t *cache)
 {
-  printf("%" PRIu64 " hits/%" PRIu64 " misses; ratio: %f\n", cache->hits, cache->misses, (float)cache->hits/(float)cache->misses);
+  printf("%" PRIu64 " hits/%" PRIu64 " misses; ratio: %f\n", cache->hits, cache->misses, (double)cache->hits/(double)cache->misses);
 }
 #endif
